+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### BUG FIXES
 
+- `[mempool]` fix `AppMempool` ABCI response handling, callback/cleanup, and lifecycle
+  ([\#5942](https://github.com/cometbft/cometbft/pull/5942))
 - `[spec]` fix the inductive invariant `spec/light-client/accountability`
 
 ### IMPROVEMENTS